refactored challenges routes and functions
