--- routes/roadmap-site/index.js
@@ -1,6 +1,0 @@
-const express = require('express')
-const app = express()
-
-app.use('/challenges/', require('./challenges'))
-
-module.exports = app
